refactoring : View를 InputView와 OutputView 2가지로 정리
