+++ src/main/java/controller/LottoStatisticsController.java
@@ -3,14 +3,14 @@ package controller;
 import model.LottoStatistics;
 import model.PurchasedLottos;
 import model.WinLotto;
-import view.LottoStatisticsView;
+import view.OutputView;
 
 public class LottoStatisticsController {
-    public LottoStatisticsView lottoStatisticsView = new LottoStatisticsView();
+    private final OutputView outputView = new OutputView();
 
     public void lottoResult(PurchasedLottos purchasedLottos, WinLotto winLotto) {
         LottoStatistics lottoStatistics = new LottoStatistics(purchasedLottos.getLottos(), winLotto);
-        lottoStatisticsView.printResult(lottoStatistics);
-        lottoStatisticsView.printTotalReturn(lottoStatistics, purchasedLottos);
+        outputView.printResult(lottoStatistics);
+        outputView.printTotalReturn(lottoStatistics, purchasedLottos);
     }
 }
