+++ src/main/java/controller/WinLottoController.java
@@ -2,16 +2,18 @@ package controller;
 
 import java.util.List;
 import model.WinLotto;
-import view.WinLottoView;
+import view.InputView;
+import view.OutputView;
 
 public class WinLottoController {
-    private WinLottoView winLottoView = new WinLottoView();
+    private final OutputView outputView = new OutputView();
+    private final InputView inputView = new InputView();
 
     public WinLotto winLotto() {
-        winLottoView.printWinNumberGuide();
-        List<Integer> winNumbers = winLottoView.readWinNumbers();
-        winLottoView.printBonusNumberGuide();
-        Integer bonusNumber = winLottoView.readBonusNumber();
+        outputView.printWinNumberGuide();
+        List<Integer> winNumbers = inputView.readWinNumbers();
+        outputView.printBonusNumberGuide();
+        Integer bonusNumber = inputView.readBonusNumber();
         return new WinLotto(winNumbers, bonusNumber);
     }
 }
