+++ src/test/java/view/InputViewTest.java
@@ -8,17 +8,17 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class PurchasedLottosViewTest {
+public class InputViewTest {
     @ParameterizedTest
     @ValueSource(strings = {"abc", "-1000", "0"})
     @DisplayName("양의정수 판별 테스트")
-    public void positiveIntegerTest(String test) {
+    public void validatePositiveNumberTest(String test) {
         // given
         System.setIn(new ByteArrayInputStream(test.getBytes()));
-        PurchaseView purchaseView = new PurchaseView();
+        InputView inputView = new InputView();
 
         // when & then
-        assertThatThrownBy(purchaseView::readPurchaseAmount)
+        assertThatThrownBy(inputView::readPurchaseAmount)
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(ErrorMessage.POSITIVE_NUMBER_EXCEPTION);
     }
